requirements.txt
playwright==1.55.0
msal==1.34.0
requests==2.32.3
python-dotenv==1.0.1
